split modules into folders
