+++ src/db/errors.rs
@@ -1,0 +1,15 @@
+use bytes::Bytes;
+
+#[derive(Debug, PartialEq, Eq)]
+pub enum Error {
+    LogDoesNotExist,
+    ItrExistsWithSameName,
+    ItrDoesNotExist,
+}
+
+impl From<Error> for Bytes {
+    fn from(e: Error) -> Self {
+        format!("{:?}", e).into()
+    }
+}
+
